Broke RegisteredTierCard out of DAR.
